Refactor entities and events; add domain and aggregate interfaces
Refactored `TestEntity` to use constructor initialization for `Id` and changed it from a record to a class. Updated `EventEntity` and `Entity` to be generic with `TKey` and `TTimeStamp` type parameters. Changed `EventVersion` type to `ulong` in `Event`, `IEvent`, and `IEventEntity`. Added new properties to `IEntity` and removed `IEntityConcurrency` and `IEntityRelational` interfaces. Introduced `EventDomain`, `IEventDomain`, `IAggregate`, and `IEventConverter` interfaces for domain events and aggregates.