Ensure default entity status and update timestamps
The `DataContext` class now ensures that new entities have a default status of `EntityStatus.ACTIVE` if their status is `null`. It also updates the last modified date of entities when their state changes to `Modified`.

The `Entity<TKey>` class introduces a new method `SetUpdatedOn` to update the `UpdatedOn` property, defaulting to the current date and time if no date is provided.

The `IEntity` interface now includes a `SetUpdatedOn` method to set the last update date, with an optional parameter for the date, defaulting to the current date and time if not provided.