Allow RequestUri to be relative or absolute
Updated the instantiation of the `Uri` object for `context.Message.RequestUri` in `HttpClientRequestQueryStringBuilder` to use `UriKind.RelativeOrAbsolute`. This change enhances flexibility by allowing the `RequestUri` to be either a relative or an absolute URI, accommodating scenarios where only a relative path is needed or when a full absolute URI is provided.