Refactor KeyId type and update comments
Updated the summary comment for AddXPipelineDeciderDecorator method in ServiceCollectionDispatcherExtensions.cs to reflect it adds a "command pipeline decorator". Modified PeekAsync call in AggregateDeciderPipelineDecorator.cs to cast request.KeyId to Guid. Changed KeyId type from Guid to object in Command<TDependency> record in Command.cs and in IDeciderCommand interface in ICommand.cs. Corrected a typo in the summary comment for IDeciderCommand interface in ICommand.cs.