Update LangVersion to preview in project files
The `<LangVersion>` property in `Xpandables.Net.AsyncPaged.csproj`
and `Xpandables.Net.Tasks.csproj` was updated from `latest` to
`preview`. This enables the use of experimental or upcoming C#
language features that are not yet part of the stable release.
This change facilitates early adoption of new features but may
introduce risks due to the potential instability of preview features.